Survey Data at Shopping Mall
The management team of a shopping mall conducted a market survey to
understand the types of people who are, or could be, visiting their mall. They are
considering designing and positioning the shopping mall services better in order to
attract mainly a few profitable market segments, or to differentiate their services
(e. invitations to events, discounts, etc ) across market